changefeedccl/kvevent: extract subpackage from kvfeed
The Event code was pretty self-contained. It feels ripe for extraction in order
to improve testing and clarify concept. The kvfeed package depends on this new
package. This commit is purely code movement.

Release note: None
