+++ pkg/ccl/changefeedccl/BUILD.bazel
@@ -29,6 +29,7 @@ go_library(
         "//pkg/ccl/backupccl/backupresolver",
         "//pkg/ccl/changefeedccl/changefeedbase",
         "//pkg/ccl/changefeedccl/changefeeddist",
+        "//pkg/ccl/changefeedccl/kvevent",
         "//pkg/ccl/changefeedccl/kvfeed",
         "//pkg/ccl/changefeedccl/schemafeed",
         "//pkg/ccl/utilccl",
@@ -129,6 +130,7 @@ go_test(
         "//pkg/blobs",
         "//pkg/ccl/changefeedccl/cdctest",
         "//pkg/ccl/changefeedccl/changefeedbase",
+        "//pkg/ccl/changefeedccl/kvevent",
         "//pkg/ccl/changefeedccl/kvfeed",
         "//pkg/ccl/changefeedccl/schemafeed",
         "//pkg/ccl/importccl",
