+++ pkg/ccl/changefeedccl/kvevent/metrics.go
@@ -6,7 +6,7 @@
 //
 //     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
 
-package kvfeed
+package kvevent
 
 import (
 	"time"
