+++ pkg/ccl/changefeedccl/kvfeed/kv_feed.go
@@ -17,6 +17,7 @@ import (
 	"fmt"
 
 	"github.com/cockroachdb/cockroach/pkg/ccl/changefeedccl/changefeedbase"
+	"github.com/cockroachdb/cockroach/pkg/ccl/changefeedccl/kvevent"
 	"github.com/cockroachdb/cockroach/pkg/ccl/changefeedccl/schemafeed"
 	"github.com/cockroachdb/cockroach/pkg/gossip"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
@@ -42,8 +43,8 @@ type Config struct {
 	Gossip             gossip.OptionalGossip
 	Spans              []roachpb.Span
 	Targets            jobspb.ChangefeedTargets
-	Sink               EventBufferWriter
-	Metrics            *Metrics
+	Sink               kvevent.Writer
+	Metrics            *kvevent.Metrics
 	MM                 *mon.BytesMonitor
 	WithDiff           bool
 	SchemaChangeEvents changefeedbase.SchemaChangeEventClass
@@ -80,8 +81,8 @@ func Run(ctx context.Context, cfg Config) error {
 		pff = rangefeedFactory(distSender.RangeFeed)
 	}
 
-	bf := func() EventBuffer {
-		return &errorWrapperEventBuffer{makeMemBuffer(cfg.MM.MakeBoundAccount(), cfg.Metrics)}
+	bf := func() kvevent.Buffer {
+		return kvevent.NewErrorWrapperEventBuffer(kvevent.NewMemBuffer(cfg.MM.MakeBoundAccount(), cfg.Metrics))
 	}
 
 	f := newKVFeed(
@@ -134,21 +135,21 @@ type kvFeed struct {
 	withDiff            bool
 	withInitialBackfill bool
 	initialHighWater    hlc.Timestamp
-	sink                EventBufferWriter
+	sink                kvevent.Writer
 	codec               keys.SQLCodec
 
 	schemaChangeEvents changefeedbase.SchemaChangeEventClass
 	schemaChangePolicy changefeedbase.SchemaChangePolicy
 
 	// These dependencies are made available for test injection.
-	bufferFactory func() EventBuffer
+	bufferFactory func() kvevent.Buffer
 	tableFeed     schemafeed.SchemaFeed
 	scanner       kvScanner
 	physicalFeed  physicalFeedFactory
 }
 
 func newKVFeed(
-	sink EventBufferWriter,
+	sink kvevent.Writer,
 	spans []roachpb.Span,
 	schemaChangeEvents changefeedbase.SchemaChangeEventClass,
 	schemaChangePolicy changefeedbase.SchemaChangePolicy,
@@ -158,7 +159,7 @@ func newKVFeed(
 	tf schemafeed.SchemaFeed,
 	sc kvScanner,
 	pff physicalFeedFactory,
-	bf func() EventBuffer,
+	bf func() kvevent.Buffer,
 ) *kvFeed {
 	return &kvFeed{
 		sink:                sink,
@@ -369,8 +370,8 @@ func (e *errBoundaryReached) Error() string {
 // *errBoundaryReached. A nil error will never be returned.
 func copyFromSourceToSinkUntilTableEvent(
 	ctx context.Context,
-	sink EventBufferWriter,
-	source EventBufferReader,
+	sink kvevent.Writer,
+	source kvevent.Reader,
 	cfg physicalConfig,
 	tables schemafeed.SchemaFeed,
 ) error {
@@ -400,7 +401,7 @@ func copyFromSourceToSinkUntilTableEvent(
 			}
 			return nil
 		}
-		applyScanBoundary = func(e Event) (skipEvent, reachedBoundary bool, err error) {
+		applyScanBoundary = func(e kvevent.Event) (skipEvent, reachedBoundary bool, err error) {
 			if scanBoundary == nil {
 				return false, false, nil
 			}
@@ -408,9 +409,9 @@ func copyFromSourceToSinkUntilTableEvent(
 				return false, false, nil
 			}
 			switch e.Type() {
-			case KVEvent:
+			case kvevent.TypeKV:
 				return true, false, nil
-			case ResolvedEvent:
+			case kvevent.TypeResolved:
 				boundaryResolvedTimestamp := scanBoundary.Timestamp().Prev()
 				resolved := e.Resolved()
 				if resolved.Timestamp.LessEq(boundaryResolvedTimestamp) {
@@ -425,11 +426,11 @@ func copyFromSourceToSinkUntilTableEvent(
 				return false, false, nil
 			}
 		}
-		addEntry = func(e Event) error {
+		addEntry = func(e kvevent.Event) error {
 			switch e.Type() {
-			case KVEvent:
+			case kvevent.TypeKV:
 				return sink.AddKV(ctx, e.KV(), e.PrevValue(), e.BackfillTimestamp())
-			case ResolvedEvent:
+			case kvevent.TypeResolved:
 				// TODO(ajwerner): technically this doesn't need to happen for most
 				// events - we just need to make sure we forward for events which are
 				// at scanBoundary.Prev(). We may not yet know about that scanBoundary.
